Ignoring package lock file.
Best practice is to not include a lock file for libraries, only for applications. I believe the idea here is that it would allow the application using the library to dictate what versions of the same dependencies should be used, as long as they fall within the same semantic version.